Split up webview_new
Cleans up the warning-logging and the webview configuration.

But this is mostly useful for isolating the :wxWebView.new error case,
which raises an exception and would otherwise be awkward to handle
along with other logic.  The next patch will address this failure case.

FIXME: Need to review the non-Edge Windows case, there might be a
behavior change.  Previously, the code would return the wxHtmlWindow
in place of a wxWebView but since the application stops here I think
it's wrong to continue as if the webview is usable.  This patch
now returns a nil webview.