Gracefully handle erlang-wx compiled without wxWebView support (#28)
* Missing webview state deprecates conditionals

If wxWebView can't be initialized, then `ui.webview` is set to `nil`.
This can be checked in functions that were previously guarded with
`webview?/0`, and the predicate function deprecated.

* Split up webview_new

Cleans up the warning-logging and the webview configuration.

But this is mostly useful for isolating the :wxWebView.new error case,
which raises an exception and would otherwise be awkward to handle
along with other logic.  The next patch will address this failure case.

FIXME: Need to review the non-Edge Windows case, there might be a
behavior change.  Previously, the code would return the wxHtmlWindow
in place of a wxWebView but since the application stops here I think
it's wrong to continue as if the webview is usable.  This patch
now returns a nil webview.

* Complain if erlang-wx exists but wxWebView is unsupported

Unfortunately, many erlang-wx distributions are lacking webview
support.  This line is the first place we can detect such a failure.

Closes #27